- rename FinanceDetails -> FinancialDetails to match field name
