+++ test/v2/fixtures/retrieveBalanceAndTransactions/RetrieveBalanceAndTransactionsResponseFixture.scala
@@ -8,26 +8,26 @@ package v2.fixtures.retrieveBalanceAndTransactions
 import play.api.libs.json.{JsValue, Json}
 import v2.models.response.retrieveBalanceAndTransactions.RetrieveBalanceAndTransactionsResponse
 
-trait RetrieveBalanceAndTransactionsResponseFixture extends BalanceDetailsFixture with DocumentDetailsFixture with FinanceDetailsFixture {
+trait RetrieveBalanceAndTransactionsResponseFixture extends BalanceDetailsFixture with DocumentDetailsFixture with FinancialDetailsFixture {
 
   val responseModel: RetrieveBalanceAndTransactionsResponse = RetrieveBalanceAndTransactionsResponse(
     balanceDetailsObject,
     Some(Seq(documentDetails)),
-    Some(Seq(financeDetailsFullObject))
+    Some(Seq(financialDetailsFullObject))
   )
 
   val responseMtdJson: JsValue = Json.parse(s"""
       |{
       |  "balanceDetails": $mtdResponseJson,
       |  "documentDetails": [$mtdDocumentDetailsJson],
-      |  "financialDetails": [$mtdFinanceDetailFullJson]
+      |  "financialDetails": [$mtdFinancialDetailsFullJson]
       |}""".stripMargin)
 
   val responseDownstreamJson: JsValue = Json.parse(s"""
        |{
        |  "balanceDetails": $downstreamResponseJson,
        |  "documentDetails": [$downstreamDocumentDetailsJson],
-       |  "financialDetails": [$downstreamFinanceDetailFullJson]
+       |  "financialDetails": [$downstreamFinancialDetailsFullJson]
        |}""".stripMargin)
 
 }
